+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## 1.4.47
 * Fixed: attempt at fixing incorrect sorting of Project roots
-* 
+* Improvement: Better type resolve for parameters without type tag in functionLiterals
 
 
 ## 1.4.46